requirements.txt
# Core
pydantic>=2.0
python-dotenv>=1.0
fastapi>=0.100
jinja2>=3.1.2
pyyaml>=6.0.1

# AI & LLM
google-generativeai>=0.3

# Storage (Future)
# google-api-python-client
